phase4/frontend/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
pg8000==1.31.0
python-dotenv==1.0.1
cohere==5.11.0
pydantic==2.10.0
python-multipart==0.0.20
psycopg2-binary==2.9.10